+++ CHANGELOG.md
@@ -29,6 +29,7 @@
 - Validate search query when overdue search query is changed
 - Bump overdue sections feature flag version to v1
 - Save overdue search history as `String` type in preferences
+- Implement `OverdueSearchHistory`
 - [In Progress: 27 Apr 2022] Migrate `BloodPressureHistoryScreenEffectHandler` to use view effect
 
 ## 2022-06-13-8291